+++ fmpz_vec/scalar_addmul_si_2exp.c
@@ -63,7 +63,7 @@ _fmpz_vec_scalar_addmul_si_2exp(fmpz * vec1, const fmpz * vec2, slong len2,
             for (i = 0; i < len2; i++)
             {
                 fmpz_mul_2exp(temp, vec2 + i, exp);
-                fmpz_submul_ui(vec1 + i, temp, -c);
+                fmpz_submul_ui(vec1 + i, temp, -((ulong) c));
             }
         }
     }
